requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
openai
faiss-cpu
tiktoken
# # uncomment to use huggingface llms
huggingface-hub

# # uncomment to use instructor embeddings
InstructorEmbedding
sentence-transformers